work done on bounding box calculations

+++ Collision.Core/Models/Position.cs
@@ -39,15 +39,15 @@ namespace Collision.Core.Models
 
         public decimal? X1 { get; set; }
         public decimal? Y1 { get; set; }
-        public int? Z1 { get; set; }
+        public decimal? Z1 { get; set; }
 
         public decimal? X2 { get; set; }
         public decimal? Y2 { get; set; }
-        public int? Z2 { get; set; }
+        public decimal? Z2 { get; set; }
 
         public decimal? X3 { get; set; }
         public decimal? Y3 { get; set; }
-        public int? Z3 { get; set; }
+        public decimal? Z3 { get; set; }
 
         public DateTime CreatedAtUtcTimeStamp { get; set; }
         public DateTime? ModifiedAtUtcTimeStamp { get; set; }
