+++ Collision.Sql.Ef/Position.cs
@@ -39,13 +39,13 @@ namespace Collision.Sql.Ef
         public Nullable<System.DateTime> UtcTimeStamp3 { get; set; }
         public Nullable<decimal> X1 { get; set; }
         public Nullable<decimal> Y1 { get; set; }
-        public Nullable<int> Z1 { get; set; }
+        public Nullable<decimal> Z1 { get; set; }
         public Nullable<decimal> X2 { get; set; }
         public Nullable<decimal> Y2 { get; set; }
-        public Nullable<int> Z2 { get; set; }
+        public Nullable<decimal> Z2 { get; set; }
         public Nullable<decimal> X3 { get; set; }
         public Nullable<decimal> Y3 { get; set; }
-        public Nullable<int> Z3 { get; set; }
+        public Nullable<decimal> Z3 { get; set; }
         public Nullable<decimal> Radius { get; set; }
         public System.DateTime CreatedAtUtcTimeStamp { get; set; }
         public Nullable<System.DateTime> ModifiedAtUtcTimeStamp { get; set; }
